termios: Workaround vet on go tip
